Rename classes as types
